# docs/apidog-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "RAG POC Module API",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "http://localhost:3000" }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": { "message": { "type": "string" } }
      },
...
      "LoginRequest": {
        "type": "object",
        "required": ["id", "password"],
        "properties": {
          "id": { "type": "string" },
          "password": { "type": "string", "format": "password" }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "Unauthorized": {
        "description": "Unauthorized",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      }
    }
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
    "/api/auth/login": { "post": { "tags": ["Auth"], "security": [], "summary": "Login", "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LoginRequest" } } } }, "responses": { "200": { "description": "OK" }, "401": { "$ref": "#/components/responses/Unauthorized" } } } },
...
    "/api/apps/{id}": { "get": { "tags": ["Apps"], "summary": "Get app", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "OK" }, "400": { "$ref": "#/components/responses/BadRequest" } } } },
...
    "/api/chat/sessions/{id}": { "get": { "tags": ["Chat Sessions"], "summary": "Get chat session", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }], "responses": { "200": { "description": "OK" }, "400": { "$ref": "#/components/responses/BadRequest" } } } },
...
  },
...
}